Final commit: numofupdates, output file (logfile_ not implemented

+++ Functions/Implementator.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-bool isPrime(int n) {
+bool isPrime(unsigned long n) {
 
     if (n <= 1){
         return false;
